--- src/core/calendar/update/index.ts
@@ -1,2 +1,0 @@
-export { updateCurrentMonth } from "./updateCurrentMonth";
-export { updateSelectedDate } from "./updateSelectedDate";
